refactor/gui_launcher
